L2CAP Support for ISO 18013-5 Holders
NOTE: Currently not working!  See below.

This provides L2CAP support for the holder for 18013-5.  This is
disabled at present; the code is in place, but calling connect() on
the L2CAP socket throws a "resources not available" exception for
reasons I have not yet tracked down.

The code should otherwise be ready, however; the request read and
the response write are both in place.  The presence of an L2CAP
characteristic on the peripheral no longer breaks the old flow for
the code, so we now properly do non-L2CAP connections to L2CAP
peripherals.

For convenience, this also includes a makefile to orchestrate
building and running; `make` or `make install` will build the
example app and attempt to install it on all connected android
devices.  `make run` will launch the example app on one device
and attempt to attach `logcat`. `make help` lists the available
commands.